+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.3] - 2026-02-27
+
+### Changes
+
+- Implement 'stats' group:
+    - Add Python/Rust library commands:
+        - `stats_get_group_data`
+        - `stats_list_groups`
+    - Add CLI commands:
+        - `stats`
+            - `get`
+            - `list-groups`
+
 ## [0.11.2] - 2026-02-16
 
 ### Changes
@@ -323,6 +336,7 @@ Initial release, not feature complete yet.
 
 Primarily to test release workflow.
 
+[0.11.3]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.11.2...0.11.3
 [0.11.2]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.11.1...0.11.2
 [0.11.1]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.11.0...0.11.1
 [0.11.0]: https://github.com/Finomnis/mcumgr-toolkit/compare/0.10.0...0.11.0